feat(modules): unify module structure, add Swagger docs, and route re-exports
