+++ src/modules/admin/admin.router.js
@@ -1,0 +1,3 @@
+import router from './admin.controller.js'
+export default router
+
