+++ src/modules/admin/index.js
@@ -1,2 +1,3 @@
+export { default as adminRouter } from './admin.router.js'
 // Export all admin module components
 export { default as adminController } from './admin.controller.js'
