+++ src/modules/common/auth/auth.router.js
@@ -1,0 +1,3 @@
+import router from './auth.controller.js'
+export default router
+
