+++ src/modules/user/user.controller.js
@@ -15,6 +15,28 @@ import {
 
 const router = Router()
 
+/**
+ * @swagger
+ * /v1/user/avatar:
+ *   post:
+ *     tags: ["Users"]
+ *     summary: Upload user avatar
+ *     security:
+ *       - cookieAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         multipart/form-data:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               avatar:
+ *                 type: string
+ *                 format: binary
+ *     responses:
+ *       200:
+ *         description: Avatar uploaded
+ */
 // User profile endpoints
 // POST /v1/user/avatar  (form-data: avatar=<file>)
 router.post('/avatar', authRequired, avatarUpload.single('avatar'), async (req, res) => {
@@ -29,6 +51,24 @@ router.post('/avatar', authRequired, avatarUpload.single('avatar'), async (req, 
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/profile/{id}:
+ *   get:
+ *     tags: ["Users"]
+ *     summary: Get user profile by id
+ *     security:
+ *       - cookieAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: string
+ *     responses:
+ *       200:
+ *         description: User profile
+ */
 // GET /v1/user/profile/:id
 router.get('/profile/:id', authRequired, async (req, res) => {
   try {
@@ -39,6 +79,24 @@ router.get('/profile/:id', authRequired, async (req, res) => {
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/profile:
+ *   put:
+ *     tags: ["Users"]
+ *     summary: Update current user profile
+ *     security:
+ *       - cookieAuth: []
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *     responses:
+ *       200:
+ *         description: Updated profile
+ */
 // PUT /v1/user/profile
 router.put('/profile', authRequired, async (req, res) => {
   try {
@@ -49,6 +107,40 @@ router.put('/profile', authRequired, async (req, res) => {
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/admin/users:
+ *   get:
+ *     tags: ["Users"]
+ *     summary: List users (admin)
+ *     security:
+ *       - cookieAuth: []
+ *     parameters:
+ *       - in: query
+ *         name: cursor
+ *         schema:
+ *           type: string
+ *           format: date-time
+ *       - in: query
+ *         name: limit
+ *         schema:
+ *           type: integer
+ *       - in: query
+ *         name: role
+ *         schema:
+ *           type: string
+ *       - in: query
+ *         name: isActive
+ *         schema:
+ *           type: boolean
+ *       - in: query
+ *         name: search
+ *         schema:
+ *           type: string
+ *     responses:
+ *       200:
+ *         description: User list
+ */
 // Admin endpoints
 // GET /v1/user/admin/users
 router.get('/admin/users', authRequired, requirePermission('view_users'), async (req, res) => {
@@ -67,6 +159,18 @@ router.get('/admin/users', authRequired, requirePermission('view_users'), async 
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/admin/stats:
+ *   get:
+ *     tags: ["Users"]
+ *     summary: User analytics (admin)
+ *     security:
+ *       - cookieAuth: []
+ *     responses:
+ *       200:
+ *         description: Stats
+ */
 // GET /v1/user/admin/stats
 router.get('/admin/stats', authRequired, requirePermission('view_analytics'), async (req, res) => {
   try {
@@ -77,6 +181,34 @@ router.get('/admin/stats', authRequired, requirePermission('view_analytics'), as
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/admin/{id}/role:
+ *   put:
+ *     tags: ["Users"]
+ *     summary: Update user role (admin)
+ *     security:
+ *       - cookieAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: string
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               role:
+ *                 type: string
+ *                 enum: [user, moderator, admin]
+ *     responses:
+ *       200:
+ *         description: Updated role
+ */
 // PUT /v1/user/admin/:id/role
 router.put('/admin/:id/role', authRequired, requireAdmin, async (req, res) => {
   try {
@@ -88,6 +220,33 @@ router.put('/admin/:id/role', authRequired, requireAdmin, async (req, res) => {
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/admin/{id}/ban:
+ *   post:
+ *     tags: ["Users"]
+ *     summary: Ban a user (admin)
+ *     security:
+ *       - cookieAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: string
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               reason:
+ *                 type: string
+ *     responses:
+ *       200:
+ *         description: Banned user
+ */
 // POST /v1/user/admin/:id/ban
 router.post('/admin/:id/ban', authRequired, requirePermission('ban_users'), async (req, res) => {
   try {
@@ -99,6 +258,24 @@ router.post('/admin/:id/ban', authRequired, requirePermission('ban_users'), asyn
   }
 })
 
+/**
+ * @swagger
+ * /v1/user/admin/{id}/unban:
+ *   post:
+ *     tags: ["Users"]
+ *     summary: Unban a user (admin)
+ *     security:
+ *       - cookieAuth: []
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         required: true
+ *         schema:
+ *           type: string
+ *     responses:
+ *       200:
+ *         description: Unbanned user
+ */
 // POST /v1/user/admin/:id/unban
 router.post('/admin/:id/unban', authRequired, requirePermission('ban_users'), async (req, res) => {
   try {
